+++ src/components/v5/common/Modals/InviteMembersModal/index.ts
@@ -1,1 +1,1 @@
-export { default } from './InviteMembersModal';
+export { default } from './InviteMemberModal';
